alert dialog

+++ app/src/main/java/org/d3if0093/assessment/ui/screen/DetailScreen.kt
@@ -69,6 +69,8 @@ fun DetailScreen(navController: NavHostController, id: Long? = null) {
     var jenis by remember { mutableStateOf("") }
     var status by remember { mutableStateOf("") }
 
+    var showDialog by remember { mutableStateOf(false) }
+
     LaunchedEffect(true){
         if (id == null) return@LaunchedEffect
         val data = viewModel.getHistori(id) ?: return@LaunchedEffect
@@ -120,6 +122,12 @@ fun DetailScreen(navController: NavHostController, id: Long? = null) {
                     }
                     if (id != null) {
                         DeleteAction {
+                            showDialog = true
+                        }
+                        DisplayAlertDialog(
+                            openDialog = showDialog,
+                            onDismissRequest = { showDialog = false }) {
+                            showDialog = false
                             viewModel.delete(id)
                             navController.popBackStack()
                         }
